refactor configuration actions
