+++ nrutil/get_configuration.go
@@ -11,13 +11,6 @@ import (
 	yaml "gopkg.in/yaml.v2"
 )
 
-// Config is configuration struct read from yaml
-type Config struct {
-	NrAdminKey        string   `yaml:"nradminkey"`
-	BasePath          string   `yaml:"basepath"`
-	SyntheticMonitors []string `yaml:"syntheticmonitors"`
-}
-
 // GetConfigurationInfo reads configuration from yaml in home folder
 func GetConfigurationInfo() (*Config, error) {
 
@@ -54,29 +47,6 @@ func (c *Config) check() error {
 		return fmt.Errorf("Configuration file check errors:\n%s", sb.String())
 	}
 	return nil
-}
-
-func createBaseYamlFile(yamlfile string) error {
-	ymlDecoded, err := yaml.Marshal(&GetBaseConfiguration().DefaultConfig)
-	if err != nil {
-		return fmt.Errorf("Unable to marshal default config. %s", err.Error())
-	}
-
-	f, err := os.Create(yamlfile)
-	if err != nil {
-		return fmt.Errorf("Unable to create yaml file in home directory: %s", err.Error())
-	}
-	defer f.Close()
-	if _, err := f.Write(ymlDecoded); err != nil {
-		return fmt.Errorf("Error writing data to yaml file in home directory: %s", err.Error())
-	}
-
-	return fmt.Errorf(`
-    The configuration yaml file was not in place.
-    A new file was created at %s.
-    Please add your New Relic User Admin key, base path and Synthetics monitor GUID's you wish to manage.
-    Once set, run this again.
-    `, yamlfile)
 }
 
 // GetConfigurationFile opens configuration file and returns a byte slice of contents, or error if not exist
